+++ server/src/api/bookings/bookings.controller.ts
@@ -2,7 +2,6 @@ import { Request, Response } from 'express';
 import AbstractController from '../../classes/AbstractController';
 import model from './bookings.model';
 import errorHandler from '/errorHandler';
-import { ObjectId } from 'mongodb';
 
 class BookingController extends AbstractController {
   constructor() {
@@ -22,7 +21,8 @@ class BookingController extends AbstractController {
   public getMine = async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
-      const response = await this.model.find({ 'user._id': new ObjectId(id) });
+      const response = await this.model.find({ 'user._id': id });
+      console.log(response);
       return res.status(200).json(response);
     } catch (error) {
       return errorHandler(error, res);